preliminary support for bundling widgets
It can be useful to include custom widgets with Grist, so they
can be used offline, or for simpler archiving, or for tighter
integration with the rest of the app. This PR includes a script
for collecting custom widget files in a structured bundle. It will
need some parallel work in Grist itself to serve the bundle.